날아간 코드 복구(캘린더(controller, service 삭제 기능)

+++ src/main/java/final_backend/Calendar/service/CalendarService.java
@@ -39,8 +39,19 @@ public class CalendarService {
         return calendarRepository.findById(scheduleId).orElse(null);
     }
 
-    public void deleteMySchedule (Long scheduleId) {
+    // 삭제 기능
+    public boolean deleteSchedule(Long scheduleId, String nickName, boolean shared) {
+        CalendarDTO calendarDTO = calendarRepository.findById(scheduleId).orElse(null);
+
+        if (calendarDTO == null) {
+            return false;
+        }
+
+        if (!calendarDTO.getWriterId().equals(nickName) && !(shared && calendarDTO.getLover() != null && calendarDTO.getLover().getNickName().equals(nickName))) {
+            return false;
+        }
         calendarRepository.deleteById(scheduleId);
+        return true;
     }
 
     public List<CalendarDTO> getScheduleByNickName(String nickname) {
